Removed code that could cause a null pointer exception. Changed type of exitStatus to Integer so that it could hold null values.

+++ src/ssh.scala
@@ -43,7 +43,7 @@ class SSH(hostname: String, keyPath: String){
     }
     
     
-    val response = new ExecuteResponse(session.getExitStatus().intValue(),
+    val response = new ExecuteResponse(session.getExitStatus(),
                                        stdout.toString,
                                        stderr.toString)
